makes RenderActiveShape reusable
